+++ src/TemperatureChart2/TemperatureChart2.js
@@ -8,7 +8,7 @@ import LineChart from 'LineChart'
 import FullScreen from 'FullScreen'
 import {
     strToDate,
-} from 'utils/Date'
+} from 'utils/DateUtils'
 
 import createChartColumns from './createChartColumns'
 import chartColumns from './chartColumns'
